welcome page에서 navigation bar 안보이게 수정
