+++ front/app/(main)/@home/welcome2/page.tsx
@@ -67,7 +67,7 @@ export default function Welcome2() {
     }, [observer]);
 
     return (
-        <>
+        <main className="relative flex h-full flex-col items-center justify-center gap-1 justify-self-stretch overflow-auto">
             <Card className="w-[30rem]">
                 <div className="flex flex-col items-center gap-[30px]">
                     <DoubleSharp width="24" height="24" />
@@ -120,6 +120,6 @@ export default function Welcome2() {
                 {/* TODO : 서버에서 닉네임이 중복되었는지, 가능한 닉네임인지 확인 */}
                 <IconArrow3 className="z-10 flex h-[48px] w-[48px] shrink-0 items-center justify-center rounded-xl bg-gray-500/80 p-3 text-gray-200/50 transition-colors duration-300 hover:bg-primary hover:text-white" />
             </Card>
-        </>
+        </main>
     );
 }
